FIX-Navbar url 정리, Meet 댓글 기능 개선

+++ src/main/resources/static/js/meets/meet_detail.js
@@ -2,6 +2,9 @@ $(document).ready(() => {
     const curUrl = window.location.href.split('/');
     const idx = curUrl[curUrl.length - 1];
     showMeetDetail(idx);
+    // 지도 기능
+    const href = '/map?id=' + idx
+    $('#go-map').attr("src", href);
 
     if (localStorage.getItem('token')) {
         $.ajaxPrefilter(function (options, originalOptions, jqXHR) {
@@ -13,35 +16,6 @@ $(document).ready(() => {
     }
 });
 
-
-function showModal(idx) {
-    $('#update-modal').modal('show');
-    console.log(idx)
-    $.ajax({
-        type: 'GET',
-        url: '/meet/api/meet' + idx,
-        dataType: 'json',
-        contentType: 'application/json; charset=utf-8',
-        success: (response) => {
-            const temp_html = `
-                <div>
-                    <div>제목: ${response.title}</div>
-                    <div>내용: ${response.content}</div>
-                    <div>주소: ${response.address}</div>
-                    <div>날짜: ${response.date}</div>
-                    <div>이미지: ${response.imgUrl}</div> 
-                </div>
-            `;
-            $('#modal-content').append(temp_html);
-        },
-        error: (err) => {
-            console.log(err);
-        }
-
-
-    })
-
-}
 
 function showMeetDetail(idx) {
     $.ajax({
@@ -74,7 +48,7 @@ function showMeetDetail(idx) {
                     <div class="contents">
                         <div class="input-group mb-3">
                             <div class="form-control content" aria-label="With textarea" id="contents_box">
-                                <img id="content-img" src="${response.imgUrl}" height="250px" width="250px"><br>
+                                <img id="content-img" src="${response.imgUrl}" alt="picture" height="250px" width="250px"><br>
                                 <span id="contents_box_span">${response.content}</span>
                             </div>
                         </div>
@@ -92,20 +66,22 @@ function showMeetDetail(idx) {
 
 function deleteMeet() {
     const id = $('#idx').val();
-    console.log(id);
-    $.ajax({
-        type: "DELETE",
-        url: "/meet/api/meet/" + id,
-        dataType: 'json',
-        contentType: 'application/json; charset=utf-8',
-        success: (response) => {
-            alert("글이 삭제 되었습니다.");
-            window.location.href = '/';
-        },
-        error: (error) => {
-            console.log(error)
-        }
-    })
+    const result = confirm("댓글을 삭제 하시겠습니까?");
+    if (result) {
+        $.ajax({
+            type: "DELETE",
+            url: "/meet/api/meet/" + id,
+            dataType: 'json',
+            contentType: 'application/json; charset=utf-8',
+            success: (response) => {
+                alert("삭제 성공!");
+                window.location.href = '/meets';
+            },
+            error: (error) => {
+                console.log(error)
+            }
+        })
+    }
 }
 
 
@@ -172,22 +148,40 @@ function saveComment() {
         success: (response) => {
             alert("댓글 저장!");
             const comment = `
-                 <div class="card mb-2">
-                 <input type="hidden" id="comment-idx" value="${response['idx']}" />
-                    <div class="card-header bg-light">
-                        <i class="fa fa-comment fa"></i> 작성자 <span id="username"></span>
-                    </div>
-                    <div class="card-body">
-                        <ul class="list-group list-group-flush">
-                            <li class="list-group-item">
-                                <div class="comment_wrote">${response['comment']}</div>
-                            </li>
-                        </ul>
-                    </div>
-                </div>
+                            <div class="card mb-2">
+                                <div class="card-header bg-light">
+                                    <i class="fa fa-comment fa"></i> 작성자: ${response.user['username']}
+                                    <input id="comment-idx" hidden value="${response['idx']}">
+                                        <div class="h-auto dropdown">
+                                          <a class="btn btn-secondary dropdown-toggle" role="button" id="dropdownMenuLink" data-bs-toggle="dropdown" aria-expanded="false">
+                                            Edit
+                                          </a>
+                                           
+                                          <ul class="dropdown-menu" aria-labelledby="dropdownMenuLink">
+                                            <li><a class="dropdown-item" onclick="showUpdateComment(${response['idx']})">수정</a></li>
+                                            <li><a class="dropdown-item" onclick="deleteComment(${response['idx']})">삭제</a></li>
+                                          </ul>
+                                        </div>
+                                </div>
+                                <div class="card-body">
+                                    <ul class="list-group list-group-flush">
+                                        <li class="list-group-item">
+                                            <div class="comment_wrote">${response.comment}</div>
+                                               <!--수정 댓글-->
+                                               
+                                             <div id="comment_input_${response['idx']}" class="d-none">
+                                                <input id="comment_upvalue_${response['idx']}" type="text"  class="form-control">
+                                                <button onclick="updateComment(${response['idx']})" type="button" class="h-auto btn btn-dark mt-3">수정</button>
+                                             </div>
+                                             
+                                        </li>
+                                    </ul>
+                                </div>
+                            </div>
                 `;
-            $('#comment_list').append(comment);
             content.val("");
+            $('#comment_list').html(comment);
+            // Todo : 내림차순 추가
         },
         error: (error) => {
             console.log(error);
